--- preprocessor/tf_utils.py
@@ -1,5 +1,0 @@
-import tensorflow as tf
-
-def normalize_img(image, label):
-    """Normalizes images: `uint8` -> `float32`."""
-    return tf.cast(image, tf.float32) / 255., label
